# services/orders-service/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Orders Service API
  version: 0.1.0
  description: Receives payment orders and publishes asynchronous payment events.
servers:
  - url: /
    description: Current Orders Service host
paths:
...
  /carts/{cartId}:
    get:
      summary: Get a cart
      operationId: getCart
      parameters:
        - name: cartId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Cart and its items returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartResponse"
        "404":
          description: Cart not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /products:
    get:
      summary: List merchant products
      operationId: listProducts
      parameters:
        - $ref: "#/components/parameters/MerchantId"
      responses:
        "200":
          description: Product list.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductResponse"
    post:
      summary: Create a product
      operationId: createProduct
      parameters:
        - $ref: "#/components/parameters/MerchantId"
        - $ref: "#/components/parameters/AuthCookie"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "201":
          description: Product created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
        "403":
          description: Admin role required.
  /categories:
    get:
      summary: List merchant categories
      operationId: listCategories
      parameters:
        - $ref: "#/components/parameters/MerchantId"
      responses:
        "200":
          description: Category list.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryResponse"
    post:
      summary: Create a category
      operationId: createCategory
      parameters:
        - $ref: "#/components/parameters/MerchantId"
        - name: X-User-Role
          in: header
          required: true
          schema:
            type: string
            enum: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequest"
      responses:
        "201":
          description: Category created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryResponse"
        "403":
          description: Admin role required.
...
  /orders:
    post:
      summary: Accept a payment order
      operationId: createOrder
      description: Validates and queues an order. Payment processing happens asynchronously.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
            example:
              orderId: order-123
              customerId: customer-456
              amount: 49.99
              currency: USD
      responses:
        "202":
          description: Order accepted and published to SQS.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AcceptedOrderResponse"
        "400":
          description: Invalid order payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Unexpected internal error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /payments/{paymentId}:
    get:
      summary: Get payment status
      operationId: getPayment
      parameters:
        - $ref: "#/components/parameters/MerchantId"
        - name: paymentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Current payment state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentResponse"
        "401":
          description: Missing merchant authentication.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Payment not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
  parameters:
    AuthCookie:
      name: Cookie
      in: header
      required: true
      schema:
        type: string
      description: HttpOnly access_token cookie issued by /auth/login.
    MerchantId:
      name: X-Merchant-Id
      in: header
      required: true
      schema:
        type: string
...
  schemas:
    CartResponse:
      type: object
      required: [cartId, userId, merchantId, currency, status, items]
      properties:
        cartId: { type: string }
        userId: { type: string }
        merchantId: { type: string }
        currency: { type: string, pattern: "^[A-Z]{3}$" }
        status: { type: string }
        items:
          type: array
          items:
            type: object
            required: [cartItemId, productId, quantity, unitAmount, totalAmount]
            properties:
              cartItemId: { type: string }
              productId: { type: string }
              quantity: { type: integer }
              unitAmount: { type: integer }
              totalAmount: { type: integer }
    CreateProductRequest:
      type: object
      additionalProperties: false
      required: [name, price, currency]
      properties:
        name:
          type: string
        sku:
          type: string
        description:
          type: string
        categoryId:
          type: string
        price:
          type: integer
          minimum: 1
        currency:
          type: string
          pattern: "^[A-Z]{3}$"
        imageUrl:
          type: string
          format: uri
        images:
          type: array
          items:
            type: string
            format: uri
        stock:
          type: integer
          minimum: 0
    ProductResponse:
      allOf:
        - $ref: "#/components/schemas/CreateProductRequest"
        - type: object
          required: [productId, merchantId, status]
          properties:
            productId:
              type: string
            merchantId:
              type: string
            status:
              type: string
    CreateCategoryRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
    CategoryResponse:
      allOf:
        - $ref: "#/components/schemas/CreateCategoryRequest"
        - type: object
          required: [categoryId, merchantId, status]
          properties:
            categoryId:
              type: string
            merchantId:
              type: string
            status:
              type: string
...
    CreateOrderRequest:
      type: object
      additionalProperties: false
      required: [orderId, customerId, amount, currency]
      properties:
        orderId:
          type: string
          pattern: "^order-[A-Za-z0-9-]+$"
        customerId:
          type: string
          pattern: "^customer-[A-Za-z0-9-]+$"
        amount:
          type: number
          exclusiveMinimum: 0
        currency:
          type: string
          pattern: "^[A-Z]{3}$"
          example: USD
    AcceptedOrderResponse:
      type: object
      required: [message, eventId]
      properties:
        message:
          type: string
          example: Order accepted
        eventId:
          type: string
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string
...
    PaymentResponse:
      type: object
      required: [paymentId, provider, status]
      properties:
        paymentId:
          type: string
        orderId:
          type: string
        provider:
          type: string
          enum: [stripe, paypal]
        status:
          type: string
          enum: [pending, approved, rejected, refunded, disputed, cancelled]
        amount:
          type: integer
        currency:
          type: string
...